refactor: :truck: move DropownMenu to components
