+++ src/components/DropdownMenu/index.ts
@@ -1,0 +1,2 @@
+export { default } from './DropdownMenu';
+export * from './DropdownMenu';
