+++ src/components/index.ts
@@ -8,6 +8,7 @@ export { default as Text } from './Text';
 export { default as Token } from './Token';
 export { default as Loader } from './Loader';
 export { default as MultiDropdown } from './MultiDropdown';
+export { default as DropdownMenu } from './DropdownMenu';
 export { default as Layout } from './Layout';
 export { default as Header } from './Header';
 export { default as Container } from './Container';
